chatbot/requirements.txt
langchain
faiss-cpu
gpt4all
python-dotenv
